feat(locking): Adds locking support, cluster locks.
Introduces a generic lock manager and Acquire/Release lock stages.
Wires acquire and release lock around cluster based operations, using
the cluster name (scoped to cloud provider/account/cluster name/region)
as the lock name.

Cluster locks are held by the executing pipeline or orchestration, or
in the case of an executing strategy, by the triggering pipeline.

There are several configuration options that can be used to
incrementally enable cluster locking to experiment with effects on
existing pipelines and user interactions:

Cluster locking is disabled completely by default, but can be enabled
with the configuration property:
```yaml
locking:
  enabled: true
```

Cluster locking will run in learning mode by default where locking
failures are non fatal and are just logged:
```yaml
locking:
  enabled: true
  learningMode: true
```

Cluster locking is not the default behaviour for an execution by
default, instead on a per execution basis cluster locking can be enabled
by adding an execution flag when submitting the pipeline or
orchestration:
```json
{
  "executionFlags": {
    "lockingEnabled": "true"
  },
  "stages": [
    "..."
  ]
}
```

Cluster locking can be made the default for all executions using a
configuration property:
```yaml
locking:
  enabled: true
  executionDefault: true
```

When cluster locking is the default for all executions, a specific
execution can still opt out of it by setting the `"lockingEnabled":
"false"` executionFlag on the pipeline or orchestration.